Ai Implemented/boosts clases added/Colliders creation in levels logic
The Ai is completally functional but can be improved. Also ive created the basic system for creating colliders, just putting all the coordinates of all the forms in a txt and the program will read them and create the physBodies. The waypoints system also follows this. And created the clas boost for the OnRoad boosts, is not implemented yet. Also do some changes and redistribution of variables and methods to parents classes so the same method is not repeated in two classes.