Fill all the different characetsr classes and improve creation racers
I put in all the different characters classes the logic so they can act like an AI or be controlled by a player using the bool isPlayer. Also think and add the logic of create different types of racers taking in account the one that the player has choosen. This logic is commented because the selectionCharacter system is still not implemented